Rename 'timelineData' module to 'data'.
