+++ src/web/modules/actions.tsx
@@ -1,6 +1,6 @@
 import { configActions } from 'web/modules/config/actions';
 import { pouchDbActions } from 'web/modules/pouchDb/actions';
-import { timelineDataActions } from 'web/modules/timelineData/actions';
+import { dataActions } from 'web/modules/data/actions';
 import { navigationActions } from 'web/modules/navigation/actions';
 import { ActionUnionFrom } from 'web/utils/redux';
 
@@ -8,7 +8,7 @@ export const actions = {
   ...configActions,
   ...navigationActions,
   ...pouchDbActions,
-  ...timelineDataActions,
+  ...dataActions,
 };
 
 // Export type describing the raw ACTION OBJECTS themselves
@@ -16,7 +16,7 @@ export type ReduxAction =
   | ActionUnionFrom<typeof configActions>
   | ActionUnionFrom<typeof navigationActions>
   | ActionUnionFrom<typeof pouchDbActions>
-  | ActionUnionFrom<typeof timelineDataActions>;
+  | ActionUnionFrom<typeof dataActions>;
 
 // Export type describing the MAP from action type names to their corresponding creator functions
 export type ReduxActions = typeof actions;
