+++ src/web/modules/navigation/reducer.tsx
@@ -3,7 +3,7 @@ import { isSameModel } from 'core/models/utils';
 import { assertExhausted } from 'server/utils/types';
 import { actions, ReduxAction } from 'web/modules/actions';
 import { ReduxState } from 'web/modules/state';
-import { getModelByUuid } from 'web/modules/timelineData/getters';
+import { getModelByUuid } from 'web/modules/data/getters';
 import { getNavigationInitState, navigationInitState, NavigationState } from 'web/modules/navigation/state';
 
 export function navigationReducer(
@@ -44,10 +44,7 @@ export function navigationReducer(
     case actions.MODEL_SELECTED_FOR_EDITING.type:
       if (state.selectedScreen !== 'BgGraphScreen') return state; // not in a relevant screen -> ignore
       let modelUuidBeingEdited = null;
-      if (
-        action.model &&
-        !isSameModel(getModelByUuid(rootState.timelineData, state.modelUuidBeingEdited), action.model)
-      ) {
+      if (action.model && !isSameModel(getModelByUuid(rootState.data, state.modelUuidBeingEdited), action.model)) {
         modelUuidBeingEdited = action.model.modelUuid;
       }
       return {
