+++ src/web/modules/state.tsx
@@ -1,18 +1,18 @@
 import { configInitState, ConfigState } from 'web/modules/config/state';
 import { pouchDbInitState, PouchDbState } from 'web/modules/pouchDb/state';
-import { timelineDataInitState, TimelineDataState } from 'web/modules/timelineData/state';
+import { dataInitState, DataState } from 'web/modules/data/state';
 import { navigationInitState, NavigationState } from 'web/modules/navigation/state';
 
 export const initReduxState: ReduxState = {
   config: configInitState,
   navigation: navigationInitState,
-  timelineData: timelineDataInitState,
+  data: dataInitState,
   pouchDb: pouchDbInitState,
 };
 
 export type ReduxState = Readonly<{
   config: ConfigState;
   navigation: NavigationState;
-  timelineData: TimelineDataState;
+  data: DataState;
   pouchDb: PouchDbState;
 }>;
